Update index selection to avoid different require calls
