+++ server.js
@@ -11,7 +11,7 @@ var authorizeService = require('./controllers/authorizeService').authorizeServic
     fs = require('fs'),
     https = require('https'),
     i18n = require('i18n-2'),
-    indexes = require('./lib/' + config.indexes.indexFile),
+    indexes = require('./lib/indexes'),
     inventorySubscription = require('./lib/inventory_subscription'),
     logger = require('./lib/logger').logger.getLogger("Server"),
     mongoose = require('mongoose'),
