+++ test/controllers/tmf-apis/catalog.js
@@ -98,8 +98,7 @@ describe('Catalog API', function() {
             './../../lib/logger': testUtils.emptyLogger,
             './../../lib/store': storeClient,
             './../../lib/rss': rssClient,
-            // './../../lib/indexes': indexes,
-            [indexesPath]: indexes,
+            './../../lib/indexes': indexes,
             './../../lib/tmfUtils': tmfUtils,
             './../../lib/utils': utils,
             'async': async
